email/password changes
I had my own email and password in the code to test if it would run. Now that it works, we can take input for the email/password so it won't save them to the file. But I think we will have to create a new Gmail account that is generic, and then we can send it from there. We can do like earlham_daily_report@gmail.com with the password as 123.abc or something like that.